+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Add `ArrayBuilder::build_arc()`
  - Add `Array::[async_]retrieve_encoded_chunk[s]`
  - Add internal `fill_array_view_with_fill_value` function
+ - Add `Array::store_metadata_opt`
 
 ### Changed
  - Support `object_store` 0.9-0.10